Tweak to handle masked / fill values in flags
Previously, there was an issue causing a TypeError when trying to
read the individual value of a flag variable if the variable had a fill
value at that position.

Instead, return that there are no flags set when the flag variable is
fill. Flags are either set, or not set! There is no inbetween, so when
the flag variable is masked, assume none of the flags are set!

Fixes get_flag and get_flags_at_index.

Traceback (most recent call last):
".../ncflag/flag_wrapper.py", line 108, in get_flag_at_index
    return (self.flags[i] & self._flag_masks[index]) ==
self._flag_values[index]
TypeError: ufunc 'bitwise_and' not supported for the input types, and
the inputs could not be safely coerced to any supported types according
to the casting rule ''safe''