Fix flag setting, more docs, get mask fn
Fix setting of flags. Described in README, but previous
method was incomplete and did not handle eg good_quality_qf
type flags that were defined to be all other flags == 0.

Previous method was to OR the value with the flag, but consider
a flag with some bits set, then trying to set good_quality_qf,
the result would be nonzero OR 0 == orignal nonzero. To fix,
must AND the NOT mask with flag before applying OR meaning.

Improve documentation in places and add function to
grab the mask associated with a meaning.