fix: set_flag, option disable zeroing on unset flags
when set_flag was called for some boolean array, the bits
targeted by flag_meaning were zerod. This is bad for flags
with no flag_masks, because it means the whole value is
zeroed if the flag_meaning is not set. This wipes out any
previous flag_meanings that were set.

So, practically speaking. Chances are, you should use
set_false=False when working with a variable that doesn't
have any flag_masks.

In the goes-r ecosystem, this is generally things like
yaw_flip_flag, for example.