Change reduce to exclude entire bit vector
Change reduce method to exclude _entire_ flag vectors (all bits) from
reduction if any bits in the exclude_mask are set.

This will be more useful, for example, if you have a collection of
flags, and know that only data associated with flags meeting certain
conditions are "averaged" (for example), then this makes it possible to
use reduce to see what the flag for the averaged data would be.

The previous behavior would not give this.